+++ CHANGELOG.md
@@ -8,3 +8,4 @@ Here you can see all changes which were applied to minestorm in the past.
 
 * Added the **command** cli command
 * Added **--message** option to *stop* and *stop-all* cli commands to specify stop message
+* Added the **[server]** option to *console* command to switch to a console at start